requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.3.4
groq==0.4.2
PyPDF2==3.0.1
python-docx==1.1.0
jinja2==3.1.3
pydantic==2.6.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1 